Move fake data generation into seeders, delete init.js
