+++ package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "start": "node index.js",
     "dev": "npx nodemon index.js",
-    "db:create": "node ./db/init.js -r 5,4 -a || node ./db/init.js -h",
+    "db:create": "npx sequelize-cli db:seed:all",
     "share": "npx ngrok http 5000",
     "test": "NODE_ENV=test jest -i --detectOpenHandles --watchAll",
     "lint": "eslint --ignore-path .gitignore ."
